common/ImageType: move to the store package
This is now only used for marshalling in the store, so keep the logic
where it is used.
